Extract out Biometric Session Manager
